Add requests, responses models; server structs; new server with warp

+++ problems/projects/tasksolver/src/input_parser/input_parser.rs
@@ -3,7 +3,7 @@ use clap::Parser;
 #[derive(Parser, Debug)]
 pub struct ServerStartArguments {
     #[arg(short = 'w', long = "workers", default_value_t = 1)]
-    pub workers_count: u64,
+    pub workers_count: usize,
     #[arg(short = 'a', long = "address", default_value = "127.0.0.1")]
     pub address: String,
     #[arg(short = 'p', long = "port", default_value = "8080")]
